refactor: ResponseHandler로 클래스 분리
응답에 대한 역할을 분리하여 ResponseHandler에서 처리하도록 refactoring
