+++ src/main/java/db/Database.java
@@ -1,7 +1,6 @@
 package db;
 
 import com.google.common.collect.Maps;
-
 import model.User;
 
 import java.util.Collection;
